ENG-93: bottom tab bar, resolve routing issues, set theme to dark
