+++ client/app/(protected)/(tabs)/arena.tsx
