+++ client/app/(protected)/(tabs)/leaderboard.tsx
