+++ client/app/(protected)/(tabs)/players.tsx
